requirements.txt
flask==3.0.0
requests==2.31.0
Pillow==10.1.0 